refactor: address pre-PR review findings for enum relocation
Pre-reviewed by 11 agents; 13 findings addressed: copy inner preset dicts so PERSONALITY_PRESETS cannot leak through proxies, make the SeniorityLevel order guard meaningful via an explicit tuple, promote type-only imports to module level, relocate seniority tests to tests/unit/hr, add get_strategic_output_default coverage, refresh API docs, regenerate the openapi TS types, and refresh the architecture-report baseline.
