+++ data/architecture_report.json
@@ -68,8 +68,8 @@
     },
     "src/synthorg/api/controllers/artifacts.py": {
       "cap": 500,
-      "loc": 498,
-      "ratio": 0.996,
+      "loc": 496,
+      "ratio": 0.992,
       "tier": "code"
     },
     "src/synthorg/api/controllers/clients.py": {
@@ -350,8 +350,8 @@
     },
     "src/synthorg/config/schema.py": {
       "cap": 500,
-      "loc": 758,
-      "ratio": 1.516,
+      "loc": 755,
+      "ratio": 1.51,
       "tier": "code"
     },
     "src/synthorg/coordination/ceremony_policy/policy_resolver.py": {
@@ -374,8 +374,8 @@
     },
     "src/synthorg/core/role_catalog.py": {
       "cap": 500,
-      "loc": 417,
-      "ratio": 0.834,
+      "loc": 414,
+      "ratio": 0.828,
       "tier": "code"
     },
     "src/synthorg/docs_engine/service.py": {
@@ -638,8 +638,8 @@
     },
     "src/synthorg/hr/hiring_service.py": {
       "cap": 500,
-      "loc": 464,
-      "ratio": 0.928,
+      "loc": 460,
+      "ratio": 0.92,
       "tier": "code"
     },
     "src/synthorg/hr/performance/models.py": {
@@ -656,8 +656,8 @@
     },
     "src/synthorg/hr/promotion/service.py": {
       "cap": 500,
-      "loc": 547,
-      "ratio": 1.094,
+      "loc": 544,
+      "ratio": 1.088,
       "tier": "code"
     },
     "src/synthorg/hr/scaling/service.py": {
@@ -806,8 +806,8 @@
     },
     "src/synthorg/memory/tools/knowledge_architect.py": {
       "cap": 500,
-      "loc": 507,
-      "ratio": 1.014,
+      "loc": 503,
+      "ratio": 1.006,
       "tier": "code"
     },
     "src/synthorg/meta/appliers/_architecture_validators.py": {
@@ -944,8 +944,8 @@
     },
     "src/synthorg/persistence/postgres/conversation_repo.py": {
       "cap": 500,
-      "loc": 468,
-      "ratio": 0.936,
+      "loc": 466,
+      "ratio": 0.932,
       "tier": "code"
     },
     "src/synthorg/persistence/postgres/escalation_repo.py": {
@@ -956,8 +956,8 @@
     },
     "src/synthorg/persistence/postgres/hr_repositories.py": {
       "cap": 500,
-      "loc": 426,
-      "ratio": 0.852,
+      "loc": 422,
+      "ratio": 0.844,
       "tier": "code"
     },
     "src/synthorg/persistence/postgres/ontology_entity_repo.py": {
@@ -968,8 +968,8 @@
     },
     "src/synthorg/persistence/postgres/org_fact_repo.py": {
       "cap": 500,
-      "loc": 487,
-      "ratio": 0.974,
+      "loc": 488,
+      "ratio": 0.976,
       "tier": "repository"
     },
     "src/synthorg/persistence/postgres/project_brain_repo.py": {
@@ -1034,8 +1034,8 @@
     },
     "src/synthorg/persistence/sqlite/hr_repositories.py": {
       "cap": 500,
-      "loc": 417,
-      "ratio": 0.834,
+      "loc": 413,
+      "ratio": 0.826,
       "tier": "code"
     },
     "src/synthorg/persistence/sqlite/ontology_entity_repo.py": {
@@ -1046,8 +1046,8 @@
     },
     "src/synthorg/persistence/sqlite/org_fact_repo.py": {
       "cap": 500,
-      "loc": 478,
-      "ratio": 0.956,
+      "loc": 475,
+      "ratio": 0.95,
       "tier": "repository"
     },
     "src/synthorg/persistence/sqlite/project_brain_repo.py": {
@@ -1260,12 +1260,6 @@
       "ratio": 0.896,
       "tier": "code"
     },
-    "src/synthorg/templates/presets.py": {
-      "cap": 500,
-      "loc": 558,
-      "ratio": 1.116,
-      "tier": "code"
-    },
     "src/synthorg/templates/renderer.py": {
       "cap": 500,
       "loc": 658,
@@ -1383,7 +1377,7 @@
     "synthorg.core.clock": 154,
     "synthorg.core.critical_errors": 471,
     "synthorg.core.domain_errors": 176,
-    "synthorg.core.enums": 493,
+    "synthorg.core.enums": 461,
     "synthorg.core.error_taxonomy": 60,
     "synthorg.core.normalization": 73,
     "synthorg.core.persistence_errors": 190,
@@ -1404,6 +1398,7 @@
     "synthorg.hr.performance.tracker": 26,
     "synthorg.hr.registry": 40,
     "synthorg.hr.scaling.models": 21,
+    "synthorg.hr.seniority": 54,
     "synthorg.hr.state": 32,
     "synthorg.hr.training.models": 23,
     "synthorg.integrations.connections.catalog": 21,
@@ -1439,6 +1434,7 @@
     "synthorg.observability.events.memory": 46,
     "synthorg.observability.events.meta": 55,
     "synthorg.observability.events.ontology": 20,
+    "synthorg.observability.events.persistence": 122,
     "synthorg.observability.events.provider": 26,
     "synthorg.observability.events.security": 63,
     "synthorg.observability.events.settings": 32,
