+++ tests/unit/templates/test_presets.py
@@ -12,10 +12,14 @@ from synthorg.core.enums import (
     DecisionMakingStyle,
     RiskTolerance,
 )
+from synthorg.hr.seniority import SeniorityLevel
+from synthorg.hr.strategy_mode import StrategicOutputMode
 from synthorg.templates.presets import (
     PERSONALITY_PRESETS,
+    STRATEGIC_OUTPUT_DEFAULTS,
     generate_auto_name,
     get_personality_preset,
+    get_strategic_output_default,
 )
 
 
@@ -217,6 +221,44 @@ class TestGetPersonalityPreset:
             custom_presets=None,
         )
         assert result["communication_style"] == "concise"
+
+
+@pytest.mark.unit
+class TestGetStrategicOutputDefault:
+    @pytest.mark.parametrize(
+        ("level", "expected"),
+        [
+            (SeniorityLevel.C_SUITE, StrategicOutputMode.ADVISOR),
+            (SeniorityLevel.VP, StrategicOutputMode.ADVISOR),
+            (SeniorityLevel.DIRECTOR, StrategicOutputMode.CONTEXT_DEPENDENT),
+        ],
+    )
+    def test_strategic_levels_map_to_defaults(
+        self, level: SeniorityLevel, expected: StrategicOutputMode
+    ) -> None:
+        assert get_strategic_output_default(level) == expected
+
+    @pytest.mark.parametrize(
+        "level",
+        [
+            SeniorityLevel.JUNIOR,
+            SeniorityLevel.MID,
+            SeniorityLevel.SENIOR,
+            SeniorityLevel.LEAD,
+            SeniorityLevel.PRINCIPAL,
+        ],
+    )
+    def test_non_strategic_levels_return_none(self, level: SeniorityLevel) -> None:
+        assert get_strategic_output_default(level) is None
+
+    def test_every_seniority_level_is_handled(self) -> None:
+        """Either a default mode or ``None`` is returned for every level."""
+        for level in SeniorityLevel:
+            result = get_strategic_output_default(level)
+            assert result is None or isinstance(result, StrategicOutputMode)
+
+    def test_defaults_keys_are_valid_seniority_levels(self) -> None:
+        assert set(STRATEGIC_OUTPUT_DEFAULTS).issubset(set(SeniorityLevel))
 
 
 @pytest.mark.unit
